Leap Year Calculator
This program determines whether a given year is a leap year based on the following rules:

A year is a leap year if it is divisible by 4.
However, if the year is divisible by 100, it is only a leap year if it is also divisible by 400.
For example, the year 2000 is a leap year (divisible by 400), but the year 1900 is not (divisible by 100 but not by 400).

Concepts Covered:

Conditional logic using nested if-else statements.
Handling user input with the Scanner class.
Understanding the rules for determining leap years based on divisibility.