Armstrong Number
This program determines whether a given integer is an Armstrong number (also known as a Narcissistic number). An Armstrong number for a 3-digit number is a number that is equal to the sum of the cubes of its digits.

Concepts Covered:

Basic control flow using a while loop.
Modular arithmetic to extract digits from a number.
Using conditional statements (if-else) to check for equality.
Handling user input with the Scanner class.