Variable Arguments
This program illustrates the use of variable arguments (varargs) in Java. The max method is defined to accept a variable number of integer arguments and determine the maximum value among them.

Concepts Covered:

Understanding and implementing variable arguments (varargs) in Java.
Handling different numbers of arguments within a method.
Iterating through an array to find the maximum value.